Fix issue of app installation with admin user
In the current implementation, devices owned by the logged in user is fetched as the device list. When an admin user logged in, even though the user has permission to view and install apps to other users devices the devices does not get fetched if the admin user is not the owner.

This resolves the above issue by fetching all devices if the user has admin permission.

Resolves product-iots#55